Rename src to proper module name
