+++ bingomaker/app/app.py
@@ -1,8 +1,7 @@
 import random
 
-from flask import Flask, current_app, render_template, request
-
 from data.persistence import TilePoolDB, tile_to_dict
+from flask import Flask, current_app, render_template, request
 from game.game import Board
 
 from . import image_routes, tilepool_routes
