+++ bingomaker/app/static/script.js
