+++ bingomaker/app/static/styles.css
