+++ bingomaker/app/tilepool_routes.py
@@ -1,6 +1,5 @@
-from flask import Blueprint, current_app, request
-
 from data.persistence import SortMethod, TilePoolDB, dict_to_tile, tile_to_dict
+from flask import Blueprint, current_app, request
 from game.game import TilePool
 
 bp = Blueprint("tilepools", __name__)
