+++ bingomaker/data/dynamodb.py
@@ -3,7 +3,6 @@ import datetime
 from uuid import uuid4
 
 import boto3
-
 from data.persistence import DBResult, SortMethod, TilePoolDB, tile_to_dict
 from game.game import Tile, TilePool
