+++ bingomaker/data/file.py
