+++ bingomaker/data/memory.py
