+++ bingomaker/data/serialization.py
@@ -3,7 +3,7 @@ from typing import Any
 
 # HACK: python pattern matching behaves weirdly
 try:
-    from src.game.game import Board, Tile, TilePool
+    from bingomaker.game.game import Board, Tile, TilePool
 except ModuleNotFoundError:
     from game.game import Board, Tile, TilePool
