+++ bingomaker/images/image_manager.py
