+++ bingomaker/images/local.py
