+++ pyproject.toml
@@ -20,7 +20,7 @@ dev-dependencies = [
 ]
 
 [tool.pytest.ini_options]
-pythonpath = [".", "src"]
+pythonpath = [".", "bingomaker"]
 testpaths = ["tests"]
 filterwarnings = ["ignore::DeprecationWarning"]
 markers = [
