+++ tests/test_serialization.py
@@ -2,7 +2,7 @@ import json
 
 from examples import example_game
 
-from src.data.serialization import BoardEncoder
+from bingomaker.data.serialization import BoardEncoder
 
 
 def test_serialize_board():
